refactor: Improve folder structure of backend/utils by creating sub-folders: fetching and scoring
